Correction of PowerGraph ref_bus type from Int to Int64

+++ src/power_graph_types.jl
@@ -3,16 +3,16 @@ using LightGraphs
 abstract type PowerGraphBase end
 
 mutable struct RadialPowerGraph <: PowerGraphBase
-    G::DiGraph # graph containing the power network 
+    G::DiGraph # graph containing the power network
     mpc::Case
-    ref_bus::Int # The id of the reference bus
+    ref_bus::Int64 # The id of the reference bus
     radial::DiGraph # The graph directed from the transmission node
 end
 
 mutable struct PowerGraph <: PowerGraphBase
     G::DiGraph
     mpc::Case
-    ref_bus::Int
+    ref_bus::Int64
 end
 
 function RadialPowerGraph()
@@ -41,7 +41,7 @@ end
 
 function read_case(mpc::Case; ignore_id::Bool=false)
     G = DiGraph(nrow(mpc.bus))
-	
+
 	for branch in eachrow(mpc.branch)
 		if ignore_id
 			edge = Edge(branch[:f_bus], branch[:t_bus])
@@ -59,5 +59,3 @@ function read_case(mpc::Case; ignore_id::Bool=false)
     end
     return G, ref_bus
 end
-
-
